Feature/dbt tasks refactor (#13217)
