+++ .gitignore
@@ -75,3 +75,4 @@ libcairo.2.dylib
 
 # setuptools-scm generated files
 src/integrations/*/**/_version.py
+*.log
